+++ appserver/src/objects/units/update.cpp
@@ -23,7 +23,7 @@ namespace Update {
 //    {' ', {0,0}}
   };
 
-  void Update_Player_Position(Game::Instance &game, int8_t &px, int8_t &py, int8_t &x, int8_t &y, Units::Species &species) {
+  void Update_Player_Position(Game::Instance &game, int &px, int &py, int &x, int &y, Units::Species &species) {
 
     std::cout << "initial player position: " << game.Get_Player().position.x << " " << game.Get_Player().position.y << std::endl;
     std::cout << "initial player location: " << game.Get_Player().location.x << " " << game.Get_Player().location.y << std::endl;
